pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gmail-mcp-server"
version = "0.1.0"
description = "A local MCP server that gives Claude native tool-level access to Gmail"
requires-python = ">=3.14"
license = "MIT"
dependencies = [
    "fastmcp>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth>=2.25.0",
]

[project.scripts]
gmail-mcp = "gmail_mcp.server:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gmail_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
